[WSW-71131][RT1060][NCP] DCP CRC32 calculation occasionally returns incorrect results
[Description]
1. Fixed race condition in DCP CRC32 calculation causing occasional incorrect
results in RT1060 NCP UART device.
2. recvfrom assert in socket_recv_task

[Root Cause]
1. DCP hardware accelerator lacked mutex protection when accessed concurrently
   by multiple tasks, leading to incorrect CRC32 calculation results
2. Missing cache coherency operations (clean/invalidate) on DCP input/output
   buffers caused stale data to be used
3. The half-closed socket (peer closed, ENOTCONN) causes lwip_select()
    to return immediately, resulting in a busy socket_recv_task loop.
    The inet close command may close the socket before socket_recv_task
    runs because uart_rx_task has higher priority.

[Fix]
1. Added mutex protection to all CRC implementations (DCP/HW/SW) to serialize
   access and prevent race conditions
2. Added DCACHE_CleanByRange before DCP operation and DCACHE_InvalidateByRange
   after DCP operation to ensure cache coherency
3. Used cache-line aligned memory allocation (OSA_MemoryAllocateAlign) with
   sufficient trailing space for safe cache operations when dcache is enabled
4. Replaced size_t with uint16_t for TLV message sizes
5. Unified queue element struct sizes to 32 bytes with explicit padding for
   better alignment and efficiency
6. Change uart rx task priority to 7.

[Testing]
Verified on RT1060-RW612 NCP UART device:
- iperf rx stress test (2000+ iterations)
- No more CRC32 calculation error issues observed

[Reviewers]
Gannicus Guo, Li Long, Hui Bai, Xinyu Hong, Gaofeng Zhang, Rex Chen

+++ ncp_examples/common/ncp_adapter/ncp_pm/ncp_pm_defs.h
@@ -103,7 +103,7 @@ typedef struct _pm_notify_state
  */
 typedef struct _ncp_pm_tx_if
 {
-    int (*send_msg)(uint32_t event, void *msg, size_t size);   /**< Send a message buffer */
+    int (*send_msg)(uint32_t event, void *msg, uint16_t size);   /**< Send a message buffer */
     void (*post_event)(uint32_t event);               /**< Send a simple event code */
 } ncp_pm_tx_if_t;
